directly calling number by clicking on it
Using HTML5-Features on mobile devices, I would suggest to embed the phone number in a "tel"-href. On mobile devices, the user can simply click on the phone number to directly call it. second point is the "itemprop" attribute to make search engines know about that data.

see http://demosthenes.info/blog/536/Adding-Phone-Numbers-To-Web-Pages-With-HTML5-and-Microdata